% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backElimination.R
\name{backElimination}
\alias{backElimination}
\title{Backwards Elimination Function}
\usage{
backElimination(Xmatrix = NULL, Yvector = NULL)
}
\arguments{
\item{Xmatrix}{Predictor Variables}

\item{Yvector}{Response Variable}
}
\description{
This function conducts backwards elimination for model selections
}
\examples{
backElimination()
}
