% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardSelection.R
\name{forwardSelection}
\alias{forwardSelection}
\title{Forward Selection Method}
\usage{
forwardSelection(Xmatrix, Yvector)
}
\arguments{
\item{Xmatrix}{Predictor Variables}

\item{Yvector}{Response Variables}
}
\description{
This Function conducts forward variable selection
}
\examples{
forwardSelection()
}
