% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmBackElimination.R
\name{glmBackElimination}
\alias{glmBackElimination}
\title{Backwards Elimination for glm Function}
\usage{
glmBackElimination(Xmatrix = NULL, Yvector = NULL, fam = "gaussian")
}
\arguments{
\item{Xmatrix}{Predictor Variables}

\item{Yvector}{Response Variable}

\item{fam}{Family of General Model, defualt gaussian}
}
\description{
This function conducts backwards elimination for model selections
}
\examples{
backElimination()
}
