% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmForwardSelection.R
\name{glmForwardSelection}
\alias{glmForwardSelection}
\title{Forward glm Selection Method}
\usage{
glmForwardSelection(Xmatrix, Yvector, fam = "gaussian")
}
\arguments{
\item{Xmatrix}{Predictor Variables}

\item{Yvector}{Response Variables}

\item{fam}{Family of glm. Default = gaussian}
}
\description{
This Function conducts forward variable selection
}
\examples{
forwardSelection()
}
