% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmSurvEstConditional.R
\name{kmSurvEstConditional}
\alias{kmSurvEstConditional}
\title{Kaplan Meier Conditional Survival Estimate FOR RIGHT CENSORED}
\usage{
kmSurvEstConditional(dataToUse, timeToMeasure, conditionalTerm)
}
\arguments{
\item{dataToUse}{The dataset used to create function}

\item{timeToMeasure}{Time at which we want to calculate survival probability}

\item{conditionalTerm}{Time used as conditional term}
}
\description{
This function calculates the KM-Estimate of Conditional Survival Function with Right Censoring Considered
}
\examples{
kmSurvEstConditional()
}
