\name{mooreFunctions-package}
\alias{mooreFunctions-package}
\alias{mooreFunctions}
\docType{package}
\title{
\packageTitle{mooreFunctions}
}
\description{
\packageDescription{mooreFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mooreFunctions}
\packageIndices{mooreFunctions}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mooreFunctions}

Maintainer: \packageMaintainer{mooreFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
