% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneKnots.R
\name{tuneKnots}
\alias{tuneKnots}
\title{Tune the `k` parameter in flexsurvsplines()}
\usage{
tuneKnots(formula, data, nKnot = 20, criterion = "AIC")
}
\arguments{
\item{formula}{Formula to be fit, response must be of type Surv() and all covariates included.}

\item{data}{Data set from which model will be built}

\item{nKnot}{Number of Knots you would like to test over. For example, default value is 20 and so we will 20 models from 0 to 19 knots.}

\item{criterion}{Information Criterion to be used. Either AIC or BIC. Default = AIC}
}
\description{
This function finds the optimal number of knots for Royston-Parmar Models
}
\examples{
 
tuneKnots(Surv(time,delta)~gender+age+race,data = kidtran,nKnot = 15, criterion = "BIC")
}
